Fix use of unsigned value in conditional that tests for negativity
Using unsigned integer with a conditional such as 'if(unsigned int i <
0)' will
always result in a true statement no matter what. This is fixed by
changing
from size_t to int64_t

+++ Source/Plugins/Sampling/SamplingFilters/NearestPointFuseRegularGrids.cpp
@@ -166,7 +166,7 @@ void NearestPointFuseRegularGrids::execute()
   size_t numRefTuples = refDims[0]*refDims[1]*refDims[2];
 
   float x, y, z;
-  size_t col, row, plane;
+  int64_t col, row, plane;
   size_t refIndex;
   size_t sampleIndex;
   size_t planeComp, rowComp;
